+++ integration_tests/src/tests/mod.rs
@@ -12,7 +12,8 @@ mod commit_post_test;
 pub use commit_post_test::*;
 mod datacap_tests;
 pub use datacap_tests::*;
-// mod evm_test;
+mod evm_test;
+pub use evm_test::*;
 // mod extend_sectors_test;
 // mod init_test;
 // mod market_miner_withdrawal_test;
