+++ test_vm/tests/evm_test.rs
@@ -1,9 +1,9 @@
-use std::sync::Arc;
-
-use ethers::prelude::abigen;
-use ethers::providers::Provider;
-use ethers::{core::types::Address as EthAddress, prelude::builders::ContractCall};
 use fil_actors_evm_shared::uints::U256;
+use fil_actors_integration_tests::tests::{
+    evm_call_test, evm_create_test, evm_init_revert_data_test, evm_staticcall_delegatecall_test,
+    id_to_eth,
+};
+use fil_actors_integration_tests::util::create_accounts;
 use fil_actors_runtime::{
     test_utils::ETHACCOUNT_ACTOR_CODE_ID, test_utils::EVM_ACTOR_CODE_ID, EAM_ACTOR_ADDR,
     EAM_ACTOR_ID,
@@ -11,25 +11,13 @@ use fil_actors_runtime::{
 use fvm_ipld_blockstore::MemoryBlockstore;
 use fvm_ipld_encoding::ipld_block::IpldBlock;
 use fvm_ipld_encoding::{strict_bytes, BytesDe, RawBytes};
+use fvm_shared::METHOD_SEND;
 use fvm_shared::{address::Address, econ::TokenAmount};
-use fvm_shared::{ActorID, METHOD_SEND};
 use num_traits::Zero;
 use serde::{Deserialize, Serialize};
-use test_vm::{util::create_accounts, TestVM, TEST_FAUCET_ADDR};
+use test_vm::{TestVM, TEST_FAUCET_ADDR};
 use vm_api::util::{apply_ok, serialize_ok};
 use vm_api::VM;
-
-// Generate a statically typed interface for the contracts.
-abigen!(Recursive, "../actors/evm/tests/contracts/Recursive.abi");
-abigen!(Factory, "../actors/evm/tests/contracts/Factory.abi");
-abigen!(FactoryChild, "../actors/evm/tests/contracts/FactoryChild.abi");
-
-fn id_to_eth(id: ActorID) -> EthAddress {
-    let mut addr = [0u8; 20];
-    addr[0] = 0xff;
-    addr[12..].copy_from_slice(&id.to_be_bytes());
-    EthAddress::from_slice(&addr)
-}
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[serde(transparent)]
@@ -39,215 +27,25 @@ struct ContractParams(#[serde(with = "strict_bytes")] pub Vec<u8>);
 fn evm_call() {
     let store = MemoryBlockstore::new();
     let v = TestVM::<MemoryBlockstore>::new_with_singletons(&store);
-
     evm_call_test(&v);
-}
-
-fn evm_call_test(v: &dyn VM) {
-    let account = create_accounts(v, 1, &TokenAmount::from_whole(10_000))[0];
-    let address = id_to_eth(account.id().unwrap());
-    let (client, _mock) = Provider::mocked();
-    let contract = Recursive::new(address, Arc::new(client));
-
-    let bytecode =
-        hex::decode(include_str!("../../actors/evm/tests/contracts/Recursive.hex")).unwrap();
-
-    let create_result = v
-        .execute_message(
-            &account,
-            &EAM_ACTOR_ADDR,
-            &TokenAmount::zero(),
-            fil_actor_eam::Method::CreateExternal as u64,
-            Some(serialize_ok(&fil_actor_eam::CreateExternalParams(bytecode))),
-        )
-        .unwrap();
-
-    assert!(
-        create_result.code.is_success(),
-        "failed to create the new actor {}",
-        create_result.message
-    );
-
-    let create_return: fil_actor_eam::CreateExternalReturn =
-        create_result.ret.unwrap().deserialize().expect("failed to decode results");
-
-    let contract_params = contract.enter().calldata().expect("should serialize");
-    let call_result = v
-        .execute_message(
-            &account,
-            &create_return.robust_address.unwrap(),
-            &TokenAmount::zero(),
-            fil_actor_evm::Method::InvokeContract as u64,
-            Some(serialize_ok(&ContractParams(contract_params.to_vec()))),
-        )
-        .unwrap();
-    assert!(call_result.code.is_success(), "failed to call the new actor {}", call_result.message);
-
-    let BytesDe(return_value) =
-        call_result.ret.unwrap().deserialize().expect("failed to deserialize results");
-    let evm_ret: u32 = contract
-        .decode_output(&contract.enter().function.name, return_value)
-        .expect("failed to decode return");
-    assert_eq!(0, evm_ret, "expected contract to return 0 on success");
 }
 
 #[test]
 fn evm_create() {
     let store = MemoryBlockstore::new();
     let v = TestVM::<MemoryBlockstore>::new_with_singletons(&store);
-
     evm_create_test(&v);
-}
-
-fn evm_create_test(v: &dyn VM) {
-    let account = create_accounts(v, 1, &TokenAmount::from_whole(10_000))[0];
-
-    let address = id_to_eth(account.id().unwrap());
-    let (client, _mock) = Provider::mocked();
-    let client = Arc::new(client);
-    let factory = Factory::new(address, client.clone());
-    let factory_child = FactoryChild::new(address, client);
-
-    let bytecode =
-        hex::decode(include_str!("../../actors/evm/tests/contracts/Lifecycle.hex")).unwrap();
-
-    let create_result = v
-        .execute_message(
-            &account,
-            &EAM_ACTOR_ADDR,
-            &TokenAmount::zero(),
-            fil_actor_eam::Method::CreateExternal as u64,
-            Some(serialize_ok(&fil_actor_eam::CreateExternalParams(bytecode))),
-        )
-        .unwrap();
-
-    assert!(
-        create_result.code.is_success(),
-        "failed to create the new actor {}",
-        create_result.message
-    );
-
-    let create_return: fil_actor_eam::CreateExternalReturn =
-        create_result.ret.unwrap().deserialize().expect("failed to decode results");
-
-    let test_func = |create_func: ContractCall<_, EthAddress>| {
-        let child_addr_eth: EthAddress = {
-            let call_params = create_func.calldata().expect("should serialize");
-            let call_result = v
-                .execute_message(
-                    &account,
-                    &create_return.robust_address.unwrap(),
-                    &TokenAmount::zero(),
-                    fil_actor_evm::Method::InvokeContract as u64,
-                    Some(serialize_ok(&ContractParams(call_params.to_vec()))),
-                )
-                .unwrap();
-            assert!(
-                call_result.code.is_success(),
-                "failed to call the new actor {}",
-                call_result.message,
-            );
-            let BytesDe(return_value) =
-                call_result.ret.unwrap().deserialize().expect("failed to deserialize results");
-            factory
-                .decode_output(&create_func.function.name, return_value)
-                .expect("failed to decode return")
-        };
-
-        let child_addr = Address::new_delegated(EAM_ACTOR_ID, &child_addr_eth.0[..]).unwrap();
-
-        // Verify the child.
-        {
-            let func = factory_child.get_value();
-            let call_params = func.calldata().expect("should serialize");
-            let call_result = v
-                .execute_message(
-                    &account,
-                    &child_addr,
-                    &TokenAmount::zero(),
-                    fil_actor_evm::Method::InvokeContract as u64,
-                    Some(serialize_ok(&ContractParams(call_params.to_vec()))),
-                )
-                .unwrap();
-            assert!(
-                call_result.code.is_success(),
-                "failed to call the new actor {}",
-                call_result.message
-            );
-            let BytesDe(return_value) =
-                call_result.ret.unwrap().deserialize().expect("failed to deserialize results");
-            let res: u32 = factory_child
-                .decode_output(&func.function.name, return_value)
-                .expect("failed to decode return");
-            assert_eq!(res, 42);
-        }
-
-        // Kill it.
-        {
-            let func = factory_child.die();
-            let call_params = func.calldata().expect("should serialize");
-            let call_result = v
-                .execute_message(
-                    &account,
-                    &child_addr,
-                    &TokenAmount::zero(),
-                    fil_actor_evm::Method::InvokeContract as u64,
-                    Some(serialize_ok(&ContractParams(call_params.to_vec()))),
-                )
-                .unwrap();
-            assert!(
-                call_result.code.is_success(),
-                "failed to call the new actor {}",
-                call_result.message
-            );
-        }
-
-        // It should now be dead.
-        {
-            let func = factory_child.get_value();
-            let call_params = func.calldata().expect("should serialize");
-            let call_result = v
-                .execute_message(
-                    &account,
-                    &child_addr,
-                    &TokenAmount::zero(),
-                    fil_actor_evm::Method::InvokeContract as u64,
-                    Some(serialize_ok(&ContractParams(call_params.to_vec()))),
-                )
-                .unwrap();
-            assert!(
-                call_result.code.is_success(),
-                "failed to call the new actor {}",
-                call_result.message
-            );
-            let BytesDe(return_value) =
-                call_result.ret.unwrap().deserialize().expect("failed to deserialize results");
-            assert!(return_value.is_empty());
-        }
-        child_addr_eth
-    };
-
-    // Test CREATE2 twice because we should be able to deploy over an existing contract.
-    let eth_addr1 = test_func(factory.create_2([0; 32], 42));
-    let eth_addr2 = test_func(factory.create_2([0; 32], 42));
-    assert_eq!(eth_addr1, eth_addr2);
-
-    // Then test create and expect two different addrs.
-    let eth_addr1 = test_func(factory.create(42));
-    let eth_addr2 = test_func(factory.create(42));
-    assert_ne!(eth_addr1, eth_addr2);
 }
 
 #[test]
 fn evm_eth_create_external() {
     let store = MemoryBlockstore::new();
     let v = TestVM::<MemoryBlockstore>::new_with_singletons(&store);
-
     evm_eth_create_external_test(&v, &v);
 }
 
 // Concrete use of TestVM is required here to run `set_actor`
-// Removing it will depend on https://github.com/filecoin-project/builtin-actors/issues/1297
+// TODO(alexytsu): Removing it will depend on https://github.com/filecoin-project/builtin-actors/issues/1297
 fn evm_eth_create_external_test(v: &dyn VM, _v_concrete: &TestVM<MemoryBlockstore>) {
     // create the EthAccount
     let eth_bits = hex_literal::hex!("FEEDFACECAFEBEEF000000000000000000000000");
@@ -633,143 +431,12 @@ fn evm_delegatecall_test(v: &dyn VM) {
 fn evm_staticcall_delegatecall() {
     let store = MemoryBlockstore::new();
     let v = TestVM::<MemoryBlockstore>::new_with_singletons(&store);
-
     evm_staticcall_delegatecall_test(&v);
-}
-
-#[allow(non_snake_case)]
-fn evm_staticcall_delegatecall_test(v: &dyn VM) {
-    // test scenarios:
-    // one hop:
-    // A -> delegatecall -> B (read) OK
-    // A -> delegatecall -> B (write) -> return (read) OK
-    // two hop with sticky staticcall:
-    // A -> staticcall -> B -> delegatecall -> C (read) OK
-    // A -> staticcall -> B -> delegatecall -> C (write) FAIL
-
-    let accounts = create_accounts(v, 3, &TokenAmount::from_whole(10_000));
-
-    let bytecode =
-        hex::decode(include_str!("../../actors/evm/tests/contracts/callvariants.hex")).unwrap();
-
-    let created: Vec<_> = accounts
-        .iter()
-        .map(|account| {
-            let create_result = v
-                .execute_message(
-                    account,
-                    &EAM_ACTOR_ADDR,
-                    &TokenAmount::zero(),
-                    fil_actor_eam::Method::CreateExternal as u64,
-                    Some(serialize_ok(&fil_actor_eam::CreateExternalParams(bytecode.clone()))),
-                )
-                .unwrap();
-
-            assert!(
-                create_result.code.is_success(),
-                "failed to create the new actor {}",
-                create_result.message
-            );
-
-            let create_return: fil_actor_eam::CreateExternalReturn =
-                create_result.ret.unwrap().deserialize().expect("failed to decode results");
-
-            // Make sure we deployed an EVM actor.
-            assert_eq!(
-                &v.actor(&Address::new_id(create_return.actor_id)).unwrap().code,
-                &*EVM_ACTOR_CODE_ID
-            );
-
-            create_return
-        })
-        .collect();
-
-    // A -> staticcall -> B -> delegatecall -> C (read) OK
-    {
-        let A_act = accounts[0];
-        let A_robust_addr = created[0].robust_address.unwrap();
-        let B = id_to_eth(created[1].actor_id);
-        let C = id_to_eth(created[2].actor_id);
-        let mut params = [0u8; 68];
-        params[3] = 11;
-        params[16..][..20].copy_from_slice(B.as_ref());
-        params[48..].copy_from_slice(C.as_ref());
-
-        let call_result = v
-            .execute_message(
-                &A_act,
-                &A_robust_addr,
-                &TokenAmount::zero(),
-                fil_actor_evm::Method::InvokeContract as u64,
-                Some(serialize_ok(&ContractParams(params.to_vec()))),
-            )
-            .unwrap();
-        assert!(
-            call_result.code.is_success(),
-            "failed to call the new actor {}",
-            call_result.message
-        );
-        let BytesDe(return_value) =
-            call_result.ret.unwrap().deserialize().expect("failed to deserialize results");
-        assert_eq!(&return_value[12..], &created[1].eth_address.0);
-    }
-
-    // A -> staticcall -> B -> delegatecall -> C (write) FAIL
-    {
-        let A_act = accounts[0];
-        let A_robust_addr = created[0].robust_address.unwrap();
-        let B = id_to_eth(created[1].actor_id);
-        let C = id_to_eth(created[2].actor_id);
-        let mut params = [0u8; 68];
-        params[3] = 13;
-        params[16..][..20].copy_from_slice(B.as_ref());
-        params[48..].copy_from_slice(C.as_ref());
-
-        let call_result = v
-            .execute_message(
-                &A_act,
-                &A_robust_addr,
-                &TokenAmount::zero(),
-                fil_actor_evm::Method::InvokeContract as u64,
-                Some(serialize_ok(&ContractParams(params.to_vec()))),
-            )
-            .unwrap();
-        assert_eq!(call_result.code.value(), 33, "static call mutation did not revert");
-    }
 }
 
 #[test]
 fn evm_init_revert_data() {
     let store = MemoryBlockstore::new();
     let v = TestVM::<MemoryBlockstore>::new_with_singletons(&store);
-
     evm_init_revert_data_test(&v);
-}
-
-fn evm_init_revert_data_test(v: &dyn VM) {
-    let account = create_accounts(v, 1, &TokenAmount::from_whole(10_000))[0];
-    let create_result = v
-        .execute_message(
-            &account,
-            &EAM_ACTOR_ADDR,
-            &TokenAmount::zero(),
-            fil_actor_eam::Method::CreateExternal as u64,
-            // init code:
-            // PUSH1 0x42; PUSH1 0x0; MSTORE;
-            // PUSH1 0x20; PUSH1 0x0; REVERT
-            Some(serialize_ok(&fil_actor_eam::CreateExternalParams(vec![
-                0x60, 0x42, 0x60, 0x00, 0x52, 0x60, 0x20, 0x60, 0x00, 0xfd,
-            ]))),
-        )
-        .unwrap();
-
-    assert!(!create_result.code.is_success(), "new actor was successfully created!");
-
-    assert!(create_result.ret.is_some(), "missing return data!");
-
-    let BytesDe(revert_data) =
-        create_result.ret.unwrap().deserialize().expect("failed to deserialize revert data");
-    let mut expected = [0u8; 32];
-    expected[31] = 0x42;
-    assert_eq!(revert_data, expected);
 }
